photoDetection/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
pillow>=10.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
